eMote: the virtual timer was changed from "ticks to expire" to "fire at this time" With a "ticks to expire" for the timers, each time a timer is placed in the comparator, every timer running is adjusted. This is fine with a few similar timers, but there are major problems (even with the real-time timer) when one timer is long and another is short. Each time the short timer is restarted, the long timer (and all others) are adjusted and our current adjustment method was off by a few microseconds.  This is because when the snapshot of our time is made to see how much to adjust, we were not taking into account the processing time between the snapshot and the actual adjustment, so our adjustment was off by a few microseconds. A few microseconds is not noticeable if we only make a few adjustments before a timer fires, but if we make hundreds or even thousands of adjustments then those microseconds add up and our long timer ends up being off enormously.

+++ MicroFrameworkPK_v4_3/DeviceCode/Include/Time_decl.h
@@ -201,7 +201,7 @@ BOOL CPU_Timer_UnInitialize(UINT16 Timer);
 
 //TODO: AnanthAtSamraksh -- check if UINT64 is right
 ////BOOL CPU_TIMER_SetCompare(UINT64 CompareValue);
-BOOL CPU_Timer_SetCompare(UINT16 Timer, UINT32 CompareValue);
+BOOL CPU_Timer_SetCompare(UINT16 Timer, UINT64 CompareValue);
 
 UINT16 CPU_Timer_GetCounter(UINT16 Timer);
 UINT16 CPU_Timer_SetCounter(UINT16 Timer, UINT32 Count);
