+++ MicroFrameworkPK_v4_3/DeviceCode/Targets/Native/STM32F10x/DeviceCode/drivers/Timer/advancedtimer/netmf_advancedtimer.h
@@ -74,7 +74,7 @@ public:
 	DeviceStatus Initialize(UINT32 Prescaler, HAL_CALLBACK_FPN ISR, void* ISR_Param);
 
 	//TODO: AnanthAtSamraksh -- is INT64 for compareValue right?
-	DeviceStatus SetCompare(UINT64 counterCorrection, UINT32 compareValue, SetCompareType scType);
+	DeviceStatus SetCompare(UINT64 counterCorrection, UINT64 compareValue, SetCompareType scType);
 
 	UINT32 GetCompare();
 	UINT32 GetCounter();
